Re-factoring and big changes.
Moving the country data from the card component to the parent (app) so that I can have access to it
more easily, so that then I can write the sorting code based on the data.

Had a lot of trouble with this, still refactoring and still got bugs to fix.