Successfuly refactored the to only need one array of countries and have that state on the parent(app).
Changed to an object oriented approach to make it work. So now every search submition that's made a new instance of country is created
and the data will be filled in and edited as needed. Wether that be by making requests to the api or by user input.
By having only one array that holds the data instead of 2 that my stop being coordinated we
have way less chance of introducing bugs into our code and unexpected behavior.

Now if just have to clean up the mess of all this changes and neat up the place a little bit. At least the code is much more elegant now. It has more class ;)