Completed MVP of project.
This is the point to where I imagined the project before conceiving it. Other ideas came along the way and
although I do want to implement other features and find the project at it's current state a bit simplistic now
this is what I set out to build so it stays at this state for now. I have other plans that wait for me :))